feat: refactor Navbar component and extract navigation logic into separate modules
